requirements.txt
litellm
matplotlib
smartpy
pydantic
PyYAML
python-dotenv
polygon-api-client
networkx
